+++ packages/@expo/cli/src/start/server/metro/MetroBundlerDevServer.ts
@@ -8,12 +8,13 @@ import { getConfig } from '@expo/config';
 import * as runtimeEnv from '@expo/env';
 import { SerialAsset } from '@expo/metro-config/build/serializer/serializerAssets';
 import chalk from 'chalk';
+import fs from 'fs';
 import { AssetData } from 'metro';
-import fetch from 'node-fetch';
 import path from 'path';
 
 import { bundleApiRoute, invalidateApiRouteCache } from './bundleApiRoutes';
 import { createRouteHandlerMiddleware } from './createServerRouteMiddleware';
+import { fetchFromMetroAsync } from './fetchFromMetro';
 import { ExpoRouterServerManifestV1, fetchManifest } from './fetchRouterManifest';
 import { instantiateMetroAsync } from './instantiateMetro';
 import { metroWatchTypeScriptFiles } from './metroWatchTypeScriptFiles';
@@ -54,6 +55,7 @@ import {
 } from '../middleware/metroOptions';
 import { prependMiddleware } from '../middleware/mutations';
 import { startTypescriptTypeGenerationAsync } from '../type-generation/startTypescriptTypeGeneration';
+import { ensureEnvironmentSupportsTLSAsync } from '../webpack/tls';
 
 export type ExpoRouterRuntimeManifest = Awaited<
   ReturnType<typeof import('expo-router/build/static/renderStaticContent').getManifest>
@@ -243,7 +245,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     const bundleUrl = new URL(devBundleUrlPathname, this.getDevServerUrl()!);
 
     // Fetch the generated HTML from our custom Metro serializer
-    const results = await fetch(bundleUrl.toString());
+    const results = await fetchFromMetroAsync(bundleUrl.toString());
 
     const txt = await results.text();
 
@@ -400,16 +402,43 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     options: BundlerStartOptions
   ): Promise<DevServerInstance> {
     options.port = await this.resolvePortAsync(options);
-    this.urlCreator = this.getUrlCreator(options);
+    this.urlCreator = this.getUrlCreator({
+      ...options,
+      location: {
+        scheme: options.https ? 'https' : 'http',
+      },
+    });
 
-    const parsedOptions = {
+    const parsedOptions: Parameters<typeof instantiateMetroAsync>[1] = {
       port: options.port,
       maxWorkers: options.maxWorkers,
       resetCache: options.resetDevServer,
     };
 
+    if (options.https) {
+      // NOTE: The `devcert` package only supports localhost. IP addresses are not supported.
+      // const hostname = new URL(this.urlCreator.constructUrl()).hostname;
+      // debug('Configuring TLS to enable HTTPS support:', hostname);
+
+      const tlsConfig = await ensureEnvironmentSupportsTLSAsync(this.projectRoot, {
+        // name: hostname,
+      }).catch((error) => {
+        Log.error(chalk.red`Error creating TLS certificates: ${error}`);
+        // Log.error(chalk.gray`- hostname: ${hostname}`);
+      });
+      if (tlsConfig) {
+        debug('Using secure server options', tlsConfig);
+        parsedOptions.secureServerOptions = {
+          key: await fs.promises.readFile(tlsConfig.keyPath),
+          cert: await fs.promises.readFile(tlsConfig.certPath),
+        };
+      }
+    }
+
+    const protocol = options.https ? 'https' : 'http';
+
     // Required for symbolication:
-    process.env.EXPO_DEV_SERVER_ORIGIN = `http://localhost:${options.port}`;
+    process.env.EXPO_DEV_SERVER_ORIGIN = `${protocol}://localhost:${options.port}`;
 
     const { metro, server, middleware, messageSocket } = await instantiateMetroAsync(
       this,
@@ -450,7 +479,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
           return this.urlCreator?.constructDevClientUrl();
         } else {
           return this.urlCreator?.constructUrl({
-            scheme: 'exp',
+            scheme: options.https ? 'exps' : 'exp',
           });
         }
       },
@@ -553,8 +582,8 @@ export class MetroBundlerDevServer extends BundlerDevServer {
         // localhost isn't always correct.
         host: 'localhost',
         // http is the only supported protocol on native.
-        url: `http://localhost:${options.port}`,
-        protocol: 'http',
+        url: `${protocol}://localhost:${options.port}`,
+        protocol,
       },
       middleware,
       messageSocket,